requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
openai>=1.50.0
httpx>=0.27.2
pydantic>=2.10.0
python-dotenv>=1.0.1
aiosqlite>=0.20.0
python-telegram-bot>=21.6
openpyxl>=3.1.5
